Move Completion to nusb::transfer
