+++ src/platform/macos_iokit/device.rs
@@ -23,9 +23,9 @@ use crate::{
         internal::{
             notify_completion, take_completed_from_queue, Idle, Notify, Pending, TransferFuture,
         },
-        Buffer, ControlIn, ControlOut, Direction, TransferError,
+        Buffer, Completion, ControlIn, ControlOut, Direction, TransferError,
     },
-    Completion, DeviceInfo, Error, MaybeFuture, Speed,
+    DeviceInfo, Error, MaybeFuture, Speed,
 };
 
 use super::{
